refactor(devframe): rename node/internal subpath to node/hub-internals
This subpath is the bridge `@devframes/hub` (and any future first-party
hub adapter) uses to reach into devframe's remote-dock token machinery
and basePath resolver. Calling it "internal" while a published package
depends on it understates the contract: rename to `hub-internals` and
document it as the stable surface for first-party hub adapters.
