+++ packages/hub/src/node/__tests__/host-docks.test.ts
@@ -3,7 +3,7 @@ import { mkdtempSync } from 'node:fs'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { REMOTE_CONNECTION_KEY } from 'devframe/constants'
-import { getInternalContext } from 'devframe/node/internal'
+import { getInternalContext } from 'devframe/node/hub-internals'
 import { describe, expect, it } from 'vitest'
 import { parseRemoteConnection } from '../../client/remote'
 import { DevframeDocksHost } from '../host-docks'
