+++ packages/hub/src/node/host-docks.ts
@@ -13,7 +13,7 @@ import type { DevframeDocksUserSettings } from '../types/settings'
 import type { DevframeHubContext } from './context'
 import { REMOTE_CONNECTION_KEY } from 'devframe/constants'
 import { createStorage } from 'devframe/node'
-import { getInternalContext } from 'devframe/node/internal'
+import { getInternalContext } from 'devframe/node/hub-internals'
 import { createEventEmitter } from 'devframe/utils/events'
 import { join } from 'pathe'
 import { DEFAULT_STATE_USER_SETTINGS } from '../constants'
