+++ tsconfig.base.json
@@ -33,8 +33,8 @@
       "devframe/node/auth": [
         "./packages/devframe/src/node/auth/index.ts"
       ],
-      "devframe/node/internal": [
-        "./packages/devframe/src/node/internal/index.ts"
+      "devframe/node/hub-internals": [
+        "./packages/devframe/src/node/hub-internals/index.ts"
       ],
       "devframe/node": [
         "./packages/devframe/src/node/index.ts"
